feat: bulletproof-react structure
