+++ client/src/app/index.tsx
@@ -1,0 +1,2 @@
+export { AppRouter } from "./router";
+export { AppProviders } from "./providers";
