+++ client/src/app/providers.tsx
@@ -1,0 +1,13 @@
+import { ThemeProvider } from "@/components/theme/theme-provider";
+
+interface AppProvidersProps {
+  children: React.ReactNode;
+}
+
+export function AppProviders({ children }: AppProvidersProps) {
+  return (
+    <ThemeProvider>
+      {children}
+    </ThemeProvider>
+  );
+}
