+++ client/src/main.tsx
@@ -1,26 +1,10 @@
 import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./index.css";
-import { RouterProvider, createRouter } from "@tanstack/react-router";
+import { AppRouter } from "./app";
 
-// Import the generated route tree
-import { routeTree } from "./routeTree.gen";
-import NotFound from "./components/common/404-not-found";
-
-// Create a new router instance
-const router = createRouter({
-  routeTree,
-  defaultNotFoundComponent: NotFound,
-});
-
-// Register the router instance for type safety
-declare module "@tanstack/react-router" {
-  interface Register {
-    router: typeof router;
-  }
-}
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
-    <RouterProvider router={router} />
+    <AppRouter />
   </StrictMode>,
 );
